revert: use loadingFinished instead, removed unnecessary functions
